pyproject.toml
[project]
name = "gyandex"
dynamic = [ "version", "dependencies" ]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)"
]
description = "organizes vast information into a structured, accessible format to simplify knowledge acquisition"
readme = "README.md"
license = "AGPL-3.0-or-later"
authors = [
    { name = "Dhruv Baldawa <dhruvbaldawa@gmail.com>", email = "dhruvbaldawa@users.noreply.github.com" }
]
requires-python = ">=3.11"

[tool.poetry]
version = "0.1.0"
package-mode = true

[project.scripts]
podgen = "gyandex.cli.podgen:main"

[tool.poetry.dependencies]
python = "^3.11,<3.14"
pandas = "^2.2.3"
notebook = "^7.2.2"
torch = "^2.5.0"
langchain = "^0.3.4"
langchain-google-genai = "^2.0.1"
python-dotenv = "^1.0.1"
ipywidgets = "^8.1.5"
langchain-google-community = {extras = ["texttospeech"], version = "^2.0.1"}
pydub = "^0.25.1"
boto3 = "^1.35.49"
feedgen = "^1.0.0"
mutagen = "^1.47.0"
rich = {extras = ["jupyter"], version = "^13.9.3"}
python-slugify = "^8.0.4"
langchain-openai = "^0.2.12"
yt-dlp = "^2024.12.13"
tenacity = "^9.0.0"
pyright = "^1.1.396"
zyphra = "^0.1.0"

[tool.poetry.group.dev.dependencies]
nbstripout = "^0.7.1"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
rust-just = "^1.36.0"
responses = "^0.25.3"
ruff = "^0.8.2"
pyright = "^1.1.390"
pytest-asyncio = "^0.25.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
extend-select = ["E501", "I"]

[tool.pyright]
include = ["gyandex"]
exclude = [
    "**/__pycache__",
    "**/*_test.py",  # @FIXME: remove this once code issues are resolved
]

defineConstant = { DEBUG = true }

reportMissingImports = "error"
reportMissingTypeStubs = false

executionEnvironments = [
    { root = "gyandex" }
]
